ARGO-223 Remove old v1 metric package Remove old v1 status msg/details packages relying on old metric package Rename metricProfilesv2 package to metricProfiles Remove old v1 unused routes
