--- app/metricProfiles/metricProfilesController.go
@@ -1,280 +1,0 @@
-/*
- * Copyright (c) 2014 GRNET S.A., SRCE, IN2P3 CNRS Computing Centre
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the
- * License. You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an "AS
- * IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- *
- * The views and conclusions contained in the software and
- * documentation are those of the authors and should not be
- * interpreted as representing official policies, either expressed
- * or implied, of either GRNET S.A., SRCE or IN2P3 CNRS Computing
- * Centre
- *
- * The work represented by this source file is partially funded by
- * the EGI-InSPIRE project through the European Commission's 7th
- * Framework Programme (contract # INFSO-RI-261323)
- */
-
-package metricProfiles
-
-import (
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"strings"
-
-	"github.com/ARGOeu/argo-web-api/utils/authentication"
-	"github.com/ARGOeu/argo-web-api/utils/config"
-	"github.com/ARGOeu/argo-web-api/utils/mongo"
-	"github.com/gorilla/mux"
-)
-
-// ListPoems lists the existing metricProfiles using a specific format which is
-// compatible with the previous api call
-func ListPoems(r *http.Request, cfg config.Config) (int, http.Header, []byte, error) {
-
-	//STANDARD DECLARATIONS START
-
-	code := http.StatusOK
-	h := http.Header{}
-	output := []byte("")
-	err := error(nil)
-	contentType := "text/xml"
-	charset := "utf-8"
-
-	//STANDARD DECLARATIONS END
-
-	session, err := mongo.OpenSession(cfg.MongoDB)
-	defer mongo.CloseSession(session)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	results := []Poem{}
-	err = mongo.Find(session, cfg.MongoDB.Db, "poem_list", nil, "p", &results)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	output, err = createPoemView(results) //Render the results into XML format
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	h.Set("Content-Type", fmt.Sprintf("%s; charset=%s", contentType, charset))
-	return code, h, output, err
-}
-
-//List the existing metricProfiles for the tenant making the request
-func List(r *http.Request, cfg config.Config) (int, http.Header, []byte, error) {
-
-	//STANDARD DECLARATIONS START
-
-	code := http.StatusOK
-	h := http.Header{}
-	output := []byte("")
-	err := error(nil)
-	contentType := "text/xml"
-	charset := "utf-8"
-
-	//STANDARD DECLARATIONS END
-
-	tenantDbConfig, err := authentication.AuthenticateTenant(r.Header, cfg)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	session, err := mongo.OpenSession(tenantDbConfig)
-	defer mongo.CloseSession(session)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	results := []MongoInterface{}
-	err = mongo.Find(session, tenantDbConfig.Db, "metric_profiles", nil, "name", &results)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	output, err = createView(results) //Render the results into XML format
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	h.Set("Content-Type", fmt.Sprintf("%s; charset=%s", contentType, charset))
-	return code, h, output, err
-}
-
-// Create a new metric profile for the tenant making the request
-func Create(r *http.Request, cfg config.Config) (int, http.Header, []byte, error) {
-	//STANDARD DECLARATIONS START
-
-	code := http.StatusOK
-	h := http.Header{}
-	output := []byte("")
-	err := error(nil)
-	contentType := "text/xml"
-	charset := "utf-8"
-
-	//STANDARD DECLARATIONS END
-
-	tenantDbConfig, err := authentication.AuthenticateTenant(r.Header, cfg)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	session, err := mongo.OpenSession(tenantDbConfig)
-	defer mongo.CloseSession(session)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	input := MongoInterface{}
-	decoder := json.NewDecoder(r.Body)
-	err = decoder.Decode(&input)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	err = mongo.Insert(session, tenantDbConfig.Db, "metric_profiles", input)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	output = []byte("Metric profile successfully inserted")
-	h.Set("Content-Type", fmt.Sprintf("%s; charset=%s", contentType, charset))
-	return code, h, output, err
-
-}
-
-// Update a metric profile for the tenant making the request
-func Update(r *http.Request, cfg config.Config) (int, http.Header, []byte, error) {
-
-	//STANDARD DECLARATIONS START
-
-	docID := mux.Vars(r)["id"]
-	code := http.StatusOK
-	h := http.Header{}
-	output := []byte("")
-	err := error(nil)
-	contentType := "text/xml"
-	charset := "utf-8"
-
-	//STANDARD DECLARATIONS END
-
-	if docID == "" {
-		docID = strings.Split(r.URL.Path, "/")[4]
-	}
-
-	tenantDbConfig, err := authentication.AuthenticateTenant(r.Header, cfg)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	session, err := mongo.OpenSession(tenantDbConfig)
-	defer mongo.CloseSession(session)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	update := MongoInterface{}
-	decoder := json.NewDecoder(r.Body)
-	err = decoder.Decode(&update)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	mongo.IdUpdate(session, tenantDbConfig.Db, "metric_profiles", docID, update)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	output = []byte("Metric profile successfully updated")
-	h.Set("Content-Type", fmt.Sprintf("%s; charset=%s", contentType, charset))
-	return code, h, output, err
-
-}
-
-// Delete a metric profile for the tenant making the request
-func Delete(r *http.Request, cfg config.Config) (int, http.Header, []byte, error) {
-
-	//STANDARD DECLARATIONS START
-
-	docID := mux.Vars(r)["id"]
-	code := http.StatusOK
-	h := http.Header{}
-	output := []byte("")
-	err := error(nil)
-	contentType := "text/xml"
-	charset := "utf-8"
-
-	//STANDARD DECLARATIONS END
-
-	if docID == "" {
-		docID = strings.Split(r.URL.Path, "/")[4]
-	}
-
-	tenantDbConfig, err := authentication.AuthenticateTenant(r.Header, cfg)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	session, err := mongo.OpenSession(tenantDbConfig)
-	defer mongo.CloseSession(session)
-
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	err = mongo.IdRemove(session, tenantDbConfig.Db, "metric_profiles", docID)
-	if err != nil {
-		code = http.StatusInternalServerError
-		return code, h, output, err
-	}
-
-	output = []byte("Metric profile successfully removed")
-	h.Set("Content-Type", fmt.Sprintf("%s; charset=%s", contentType, charset))
-	return code, h, output, err
-}
